+++ dyndnsc/tests/detector/tests.py
@@ -27,7 +27,7 @@ class IndividualDetectorTests(unittest.TestCase):
         NAME = "dns"
         self.assertEqual(NAME, dyndnsc.detector.dns.IPDetector_DNS.getName())
         detector = dyndnsc.detector.dns.IPDetector_DNS("localhost")
-        self.assertFalse(detector.canDetectOffline())
+        self.assertFalse(detector.can_detect_offline())
         self.assertEqual(NAME, detector.getName())
         self.assertEqual(None, detector.getCurrentValue())
         self.assertTrue(type(detector.detect()) in (type(None), str))
@@ -40,7 +40,7 @@ class IndividualDetectorTests(unittest.TestCase):
         cmd = "echo 127.0.0.1"
         self.assertEqual(NAME, dyndnsc.detector.command.IPDetector_Command.getName())
         detector = dyndnsc.detector.command.IPDetector_Command({"command": cmd})
-        self.assertFalse(detector.canDetectOffline())
+        self.assertFalse(detector.can_detect_offline())
         self.assertEqual(NAME, detector.getName())
         self.assertEqual(None, detector.getCurrentValue())
         self.assertTrue(type(detector.detect()) in (type(None), str))
@@ -66,7 +66,7 @@ class IndividualDetectorTests(unittest.TestCase):
         NAME = "random"
         self.assertEqual(NAME, dyndnsc.detector.rand.IPDetector_Random.getName())
         detector = dyndnsc.detector.rand.IPDetector_Random()
-        self.assertTrue(detector.canDetectOffline())
+        self.assertTrue(detector.can_detect_offline())
         self.assertEqual(NAME, detector.getName())
         self.assertEqual(None, detector.getCurrentValue())
         self.assertTrue(type(detector.detect()) in (str,))
@@ -76,7 +76,7 @@ class IndividualDetectorTests(unittest.TestCase):
         NAME = "iface"
         self.assertEqual(NAME, dyndnsc.detector.iface.IPDetector_Iface.getName())
         detector = dyndnsc.detector.iface.IPDetector_Iface()
-        self.assertTrue(detector.canDetectOffline())
+        self.assertTrue(detector.can_detect_offline())
         self.assertEqual(NAME, detector.getName())
         self.assertEqual(None, detector.getCurrentValue())
         self.assertTrue(type(detector.detect()) in (type(None), str))
@@ -86,7 +86,7 @@ class IndividualDetectorTests(unittest.TestCase):
         NAME = "teredo"
         self.assertEqual(NAME, dyndnsc.detector.teredo.IPDetector_Teredo.getName())
         detector = dyndnsc.detector.teredo.IPDetector_Teredo()
-        self.assertTrue(detector.canDetectOffline())
+        self.assertTrue(detector.can_detect_offline())
         self.assertEqual(NAME, detector.getName())
         self.assertEqual(None, detector.getCurrentValue())
         self.assertTrue(type(detector.detect()) in (type(None), str))
@@ -110,7 +110,7 @@ class IndividualDetectorTests(unittest.TestCase):
         NAME = "webcheck"
         self.assertEqual(NAME, dyndnsc.detector.webcheck.IPDetector_WebCheck.getName())
         detector = dyndnsc.detector.webcheck.IPDetector_WebCheck()
-        self.assertFalse(detector.canDetectOffline())
+        self.assertFalse(detector.can_detect_offline())
         self.assertEqual(NAME, detector.getName())
         self.assertEqual(None, detector.getCurrentValue())
         self.assertTrue(type(detector.detect()) in (type(None), str))